fix: verificação do id (req.params

+++ src/controllers/wifiController.ts
@@ -27,6 +27,8 @@ export async function getWifis(req: Request, res: Response){
 export async function getWifiById(req: Request, res: Response){
     const { id } = req.params;
     const session: Session = res.locals.session;
+
+    if(!id) return res.status(400).send("Missing id");
     const user = await userRepository.findUserById(session.userId);
 
     const wifi = await wifiService.getWifiById(Number(id), user, session);
@@ -36,6 +38,8 @@ export async function getWifiById(req: Request, res: Response){
 export async function deleteWifiById(req: Request, res: Response){
     const { id } = req.params;
     const session: Session = res.locals.session;
+
+    if(!id) return res.status(400).send("Missing id");
     const user = await userRepository.findUserById(session.userId);
 
     await wifiService.deleteWifiById(Number(id), user, session);
